pyproject.toml
[project]
name = "oaps"
version = "0.0.0"
description = "An overengineered agentic project system."
readme = "README.md"
authors = [{ name = "Tony Burns", email = "tony@tonyburns.net" }]
requires-python = ">=3.14"
dependencies = [
  "anthropic>=0.75.0",
  "anyio>=4.12.0",
  "bokeh>=3.8.1",
  "claude-agent-sdk>=0.1.15",
  "cyclopts>=4.3.0",
  "dulwich>=0.24.10",
  "fastapi[standard]>=0.125.0",
  "jinja2>=3.1.6",
  "jsonschema>=4.25.1",
  "matplotlib>=3.10.0",
  "mdformat>=0.7.22",
  "mdformat-frontmatter>=2.0.8",
  "mdformat-gfm>=1.0.0",
  "mdformat-gfm-alerts>=2.0.0",
  "mdformat-ruff>=0.1.3",
  "mdformat-simple-breaks>=0.0.1",
  "orjson>=3.11.5",
  "pathspec>=0.12.1",
  "pendulum>=3.1.0",
  "platformdirs>=4.5.1",
  "polars>=1.36.1",
  "pyarrow>=22.0.0",
  "pydantic>=2.12.5",
  "pydantic-extra-types[pendulum]>=2.10.6",
  "pydantic-settings>=2.12.0",
  "pytablewriter>=1.2.1",
  "pytest>=9.0.2",
  "pyyaml>=6.0.3",
  "rich>=14.2.0",
  "rule-engine>=4.5.3",
  "rustworkx>=0.17.1",
  "seaborn>=0.13.2",
  "statsmodels>=0.14.6",
  "strawberry-graphql[cli]>=0.287.3",
  "structlog>=25.5.0",
  "tabulate>=0.9.0",
  "tenacity>=9.1.2",
  "textual>=6.8.0",
  "tomli-w>=1.2.0",
  "uvicorn>=0.38.0",
  "voyageai>=0.3.7",
  "watchfiles>=1.1.1",
  "zensical>=0.0.11",
]

[project.scripts]
oaps = "oaps.cli:main"
oaps-hook = "oaps.hooks.cli:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "basedpyright>=1.36.1",
  "codespell>=2.4.1",
  "cosmic-ray>=8.4.3",
  "cyclonedx-py>=1.0.1",
  "djlint>=1.36.4",
  "hypothesis>=6.148.7",
  "memray>=1.19.1",
  "prek>=0.2.21",
  "py-spy>=0.4.1",
  "pyfakefs>=5.10.2",
  "pytest>=9.0.2",
  "pytest-benchmark>=5.2.3",
  "pytest-cov>=7.0.0",
  "pytest-examples>=0.0.18",
  "pytest-mock>=3.15.1",
  "ruff>=0.14.9",
  "ty>=0.0.3",
  "yamllint>=1.37.1",
  "zensical>=0.0.11",
]

[tool.basedpyright]
exclude = ["**/node_modules", "**/tmp", "examples/**", ".oaps/**"]
include = ["src", "tests"]
pythonVersion = "3.14"
reportImportCycles = "error"
allowedUntypedLibraries = ["pyfakefs", "rule_engine", "unittest.mock", "zensical"]

[[tool.basedpyright.executionEnvironments]]
root = "tests"
extraPaths = ["."]
reportUnusedParameter = "none"
reportPrivateUsage = "none"
reportUnusedCallResult = "none"

[tool.codespell]
skip = "coverage.xml,coverage.json,uv.lock,htmlcov,tmp,pnpm-lock.yaml,docs,site"
ignore-words-list = "AKS,CrossReference,Nd,Wont,WONT"

[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib",
  "--tb",
  "short",
  "--ignore=tests/benchmarks",
  "-m",
  "not claude_integration",
]
pythonpath = ["src", "."]
filterwarnings = ["error"]
testpaths = ["tests"]
strict = true  # pytest 9.0+: enables strict_config, strict_markers, strict_xfail, strict_parametrization_ids
markers = [
  "benchmark: mark a test as a benchmark test.",
  "claude_integration: mark a test requiring the Claude CLI (slow, excluded by default).",
  "docexamples: mark a test as a documentation example test.",
  "integration: mark a test as an integration test.",
  "property: mark a test as a property test.",
  "unit: mark a test as a unit test.",
  # Domain markers
  "cli: mark a test as a CLI test.",
  "config: mark a test as a configuration test.",
  "core: mark a test as a core module test.",
  "hooks: mark a test as a hooks module test.",
  "server: mark a test as a server module test.",
]

[tool.coverage.run]
branch = true
source = ["src/oaps"]
relative_files = true
omit = ["tests/*", "**/__pycache__/*", "**/site-packages/*", ".oaps/**/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "@abc.abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
extend-exclude = ["**/node_modules/**", "**/tmp/**", "examples/**", ".oaps/**", "skills/**/scripts/**"]
force-exclude = true
target-version = "py314"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  "A",     # flake8-builtins
  "ARG",   # flake8-unused-arguments
  "ASYNC", # flake8-async
  "B",     # bugbear
  "BLE",   # flake8-blind-except
  "C4",    # flake8-comprehensions
  "D",     # pydocstyle
  "DTZ",   # flake8-datetimez
  "E",     # pycodestyle (error)
  "EM",    # flake8-errmsg
  "EXE",   # flake8-executable
  "F",     # pyflakes
  "FBT",   # flake8-boolean-trap
  "FURB",  # refurb
  "G",     # flake8-logging-format
  "I",     # isort
  "ICN",   # flake8-import-conventions
  "ISC",   # flake8-implicit-str-concat
  "LOG",   # flake8-logging
  "N",     # pep8-naming
  "PERF",  # Perflint
  "PGH",   # pygrep-hooks
  "PIE",   # flake8-pie
  "PL",    # pylint
  "PT",    # flake8-pytest-style
  "PTH",   # flake8-use-pathlib
  "PYI",   # flake8-pyi
  "RET",   # flake8-return
  "RUF",   # Ruff-specific rules
  "S",     # flake8-bandit
  "SIM",   # flake8-simplify
  "SLF",   # flake8-self
  "SLOT",  # flake8-slots
  "TC",    # flake8-type-checking
  "TD",    # flake8-todos
  "TID",   # flake8-tidy-imports
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "TRY",   # tryceratops
  "UP",    # pyupgrade
]
ignore = ["TD003"]
pydocstyle = { convention = "google" }

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
forced-separate = ["tests"]

[tool.ruff.lint.flake8-type-checking]
quote-annotations = true

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
  "D",       # no docstrings in tests
  "FBT001",  # boolean positional args in function defs
  "FBT003",  # boolean positional args in function calls
  "PLC0415", # function-level imports in tests (needed for pyfakefs)
  "PT012",   # pytest.raises with multiple statements (needed for context manager tests)
  "S101",    # assert statements in tests
  "S607",    # starting a process with partial executable path (common in tests)
  "SLF001",  # private member access in tests (monkeypatch internals)
  "PLR0913", # too many arguments in tests
  "PLR2004", # magic values in assertions in tests
  "ARG002",  # unused arguments in tests
  "N811",    # import alias naming in tests
  "TC",      # type-checking imports used at runtime in tests for fixtures
]
"src/oaps/commands/**" = [
  "PLC0415", # Allow function-level import statements in commands
  "T201",    # Allow print in command outputs
]
"src/oaps/cli/_commands/**" = [
  "PLC0415", # Allow function-level import statements in commands
  "T201",    # Allow print in command outputs
]
"examples/**" = ["T201"] # Allow print in examples
"scripts/**" = ["T201"] # Allow print in scripts
"src/oaps/idea/_models.py" = ["TC003"] # datetime needed at runtime for slots dataclass
"src/oaps/spec/_models.py" = ["TC003"] # datetime needed at runtime for slots dataclass
"src/oaps/idea/_manager.py" = ["PLR0912", "PLR0913"] # complex search logic, factory method
"src/oaps/hooks/specs.py" = ["PLR0911", "PLR0912", "PLR2004", "SIM105"] # complex hook processing with many validation branches
